chore(cli): moved constants to utils
moved constants to utils directory
