+++ bin/config/config-yargs.js
@@ -1,5 +1,18 @@
 const optionsSchema = require("../config/optionsSchema.json");
 
+const { GROUPS } = require("../utils/constants");
+
+const {
+	CONFIG_GROUP,
+	BASIC_GROUP,
+	MODULE_GROUP,
+	OUTPUT_GROUP,
+	ADVANCED_GROUP,
+	RESOLVE_GROUP,
+	OPTIMIZE_GROUP,
+	DISPLAY_GROUP
+} = GROUPS;
+
 const nestedProperties = ["anyOf", "oneOf", "allOf"];
 
 const resolveSchema = schema => {
@@ -51,19 +64,6 @@ const getSchemaInfo = (path, property, subProperty) => {
 	}
 	return findPropertyInSchema(current, property, subProperty);
 };
-
-const { GROUPS } = require("../constants");
-
-const {
-	CONFIG_GROUP,
-	BASIC_GROUP,
-	MODULE_GROUP,
-	OUTPUT_GROUP,
-	ADVANCED_GROUP,
-	RESOLVE_GROUP,
-	OPTIMIZE_GROUP,
-	DISPLAY_GROUP
-} = GROUPS;
 
 module.exports = function(yargs) {
 	yargs
